+++ src/lib/Navbar.svelte
@@ -167,7 +167,6 @@
 />
 
 <SettingsDialog bind:isOpen={settingsOpen} bind:settings />
-
 <div
   class="absolute top-0 left-0 w-full bg-neutral-50 dark:bg-neutral-900 shadow-md flex flex-row justify-between items-center px-6 py-4 border-b-[0.75px] border-[#b300e6]"
 >
@@ -726,7 +725,11 @@
       </a>
 
       <!-- Settings button -->
-      <button title="Open Settings" on:click={() => (settingsOpen = true)}>
+      <button
+        title="Open Settings"
+        on:click={() => (settingsOpen = true)}
+        aria-label="Open Settings"
+      >
         <svg
           xmlns="http://www.w3.org/2000/svg"
           width="24"
@@ -737,6 +740,7 @@
           stroke-width="2"
           stroke-linecap="round"
           stroke-linejoin="round"
+          class="size-6 text-white dark:text-white hover:text-neutral-900 dark:hover:text-white transition-colors"
           ><circle cx="12" cy="12" r="3"></circle><path
             d="M19.4 15a1.65 1.65 0 0 0 .33 1.82l.06.06a2 2 0 0 1 0 2.83 2 2 0 0 1-2.83 0l-.06-.06a1.65 1.65 0 0 0-1.82-.33 1.65 1.65 0 0 0-1 1.51V21a2 2 0 0 1-2 2 2 2 0 0 1-2-2v-.09A1.65 1.65 0 0 0 9 19.4a1.65 1.65 0 0 0-1.82.33l-.06.06a2 2 0 0 1-2.83 0 2 2 0 0 1 0-2.83l.06-.06a1.65 1.65 0 0 0 .33-1.82 1.65 1.65 0 0 0-1.51-1H3a2 2 0 0 1-2-2 2 2 0 0 1 2-2h.09A1.65 1.65 0 0 0 4.6 9a1.65 1.65 0 0 0-.33-1.82l-.06-.06a2 2 0 0 1 0-2.83 2 2 0 0 1 2.83 0l.06.06a1.65 1.65 0 0 0 1.82.33H9a1.65 1.65 0 0 0 1-1.51V3a2 2 0 0 1 2-2 2 2 0 0 1 2 2v.09a1.65 1.65 0 0 0 1 1.51 1.65 1.65 0 0 0 1.82-.33l.06-.06a2 2 0 0 1 2.83 0 2 2 0 0 1 0 2.83l-.06.06a1.65 1.65 0 0 0-.33 1.82V9a1.65 1.65 0 0 0 1.51 1H21a2 2 0 0 1 2 2 2 2 0 0 1-2 2h-.09a1.65 1.65 0 0 0-1.51 1z"
           ></path></svg